Consolidate duplicate directories
