+++ docs/directory_overview.md
@@ -11,9 +11,8 @@ This document describes the purpose of the main folders in the repository to mak
 
 ## Utilities and Scripts
 
-- `tools/` - Assorted helper scripts used in development and operations.
-- `scripts/` - Automation utilities such as log analysis and release prep.
-- `examples/` - Small code samples demonstrating library usage.
+- `tools/` - Assorted helper and automation scripts used in development and operations.
+- `docs/examples/` - Small code samples demonstrating library usage.
 - `docs/` - Project documentation.
 
 ## Data and Runtime
@@ -21,7 +20,6 @@ This document describes the purpose of the main folders in the repository to mak
 - `config/` - Configuration templates and environment files.
 - `runtime/` - Temporary runtime files (queues, logs, etc.).
 - `data/` - Data artifacts including dev logs.
-- `reports/` - Generated reports and output files.
 
 ## Testing
 
@@ -32,7 +30,6 @@ This document describes the purpose of the main folders in the repository to mak
 
 - `gui/` - Simple GUI components used by the project.
 - `social/` - Social media automation code.
-- `episodes/` - Sample content loops and scenarios.
 - `prompts/` - Prompt templates and examples.
 
 For a quick visual overview you can run:
