+++ dreamos/core/self_discovery/journal.py
