+++ tools/analyze_logs.py
