+++ tools/calibrate_coordinates.py
