+++ tools/create_minimal_core.py
