+++ tools/zip_resolver.py
